Added Authentication And Authorization

+++ src/main/java/com/saketmungsemajorproject/Airbnb/App/service/impl/RoomServiceImpl.java
@@ -3,7 +3,9 @@ package com.saketmungsemajorproject.Airbnb.App.service.impl;
 import com.saketmungsemajorproject.Airbnb.App.dto.RoomDto;
 import com.saketmungsemajorproject.Airbnb.App.entity.Hotel;
 import com.saketmungsemajorproject.Airbnb.App.entity.Room;
+import com.saketmungsemajorproject.Airbnb.App.entity.User;
 import com.saketmungsemajorproject.Airbnb.App.exception.ResourceNotFoundException;
+import com.saketmungsemajorproject.Airbnb.App.exception.UnAuthorisedException;
 import com.saketmungsemajorproject.Airbnb.App.repository.HotelRepository;
 import com.saketmungsemajorproject.Airbnb.App.repository.RoomRepository;
 import com.saketmungsemajorproject.Airbnb.App.service.InventoryService;
@@ -12,6 +14,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,6 +35,12 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel with Id "+hotelId+" not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         Room newRoom =  modelMapper.map(roomDto, Room.class);
         newRoom.setHotel(hotel);
         newRoom = roomRepository.save(newRoom);
@@ -48,6 +57,12 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel with Id "+hotelId+" not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
 //        List<Room> rooms = roomRepository.findAllRoomsByHotelId(hotelId);
 //        List<RoomDto> ans = rooms
 //                .stream().map((element) -> modelMapper.map(element, RoomDto.class))
@@ -75,6 +90,12 @@ public class RoomServiceImpl implements RoomService{
         Room room = roomRepository
                 .findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Room with Id "+id+" not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+room.getHotel().getId());
+        }
+
         modelMapper.map(roomDto,room);
         room.setId(id);
         room = roomRepository.save(room);
@@ -87,6 +108,11 @@ public class RoomServiceImpl implements RoomService{
         Room room = roomRepository
                 .findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Room with Id "+id+" not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+room.getHotel().getId());
+        }
 
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(id);
